Move Command resolvers to manager level
