+++ java/client/src/main/java/glide/api/ClusterClient.java
@@ -1,0 +1,2 @@
+package glide.api;public class ClusterClient {
+}
