+++ java/client/src/main/java/glide/managers/BaseCommandResponseResolver.java
@@ -1,4 +1,4 @@
-package glide.api.commands;
+package glide.managers;
 
 import glide.api.models.exceptions.ClosingException;
 import glide.api.models.exceptions.ConnectionException;
