+++ java/client/src/test/java/glide/managers/CommandManagerTest.java
@@ -9,7 +9,6 @@ import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import glide.api.commands.BaseCommandResponseResolver;
 import glide.api.commands.Command;
 import glide.api.models.exceptions.ClosingException;
 import glide.api.models.exceptions.ConnectionException;
